safari-600.7-branch/LayoutTests/fast/frames/frame-with-noresize-can-be-resized-after-removal-of-noresize-expected.txt



--------
Frame: 'testFrame'
--------


--------
Frame: 'results'
--------
This tests that a frame with a source code-specified noresize attribute can be resized after programmatically removing its noresize attribute.

PASS document.getElementById("testFrame").width is 210
